requirements.txt
pymongo
dnspython
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
pymongo
dnspython
python-dotenv
pyjwt
bcrypt
